AI_Enhanced_Cybersecurity_Project/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pytest
